+++ src/is-ios-webview.ts
@@ -9,14 +9,8 @@ export = function isIosWebview(ua?: string): boolean {
     if (isIosGoogleSearchApp(ua)) {
       return true;
     }
-
-    return (
-      // Historically, a webview could be identified by the presence of AppleWebKit and _no_ presence of Safari after.
-      /.+AppleWebKit(?!.*Safari)/i.test(ua) ||
-      // A webview can now have 'Safari', but if it does, then we want to check the `safari` window prop as webviews won't have it.
-      (ua.indexOf("Safari") > -1 &&
-        (!window.safari || !window.safari.pushNotification))
-    );
+    // Historically, a webview could be identified by the presence of AppleWebKit and _no_ presence of Safari after.
+    return /.+AppleWebKit(?!.*Safari)/i.test(ua);
   }
 
   return false;